Refactor overseas notification
